Add a simple --unset option to unset the NOTIFY_SOCKET env var.
For my usecase, I would like to use this to proxy a systemd-nspawn
container. The unit starting this container will have Type=notify and
thus expect generate a NOTIFY_SOCKET variable for the proxy, but when
starting systemd-nspawn I wish to disable the built in notifications
and instead do this in a custom unit inside the booted container which
will be an internal watchdog and thus can publish WATCHDOG=1 variables
back to the host.

This is a similar technique as documented in
  https://github.com/systemd/systemd/issues/997
and
  https://github.com/systemd/systemd/issues/1369
and by unsetting the NOTIFY_SOCKET variable we can prevent systemd-nspawn
itself from notifying systemd of the readiness state.

Combine this with an appropriate bind mount when starting the container
and we can create a custom unit to handle this notification, both of
readiness and status updates during runtime and watchdog status.